# Conflicts:
#	docs/blog/2018-04-04-gatsby-contentful-starter-tutorial/index.md
#	docs/blog/2020-01-08-git-workflows/index.md

+++ packages/gatsby/src/utils/page-data.js
@@ -15,7 +15,10 @@ const read = async ({ publicDir }, pagePath) => {
 
 const remove = async ({ publicDir }, pagePath) => {
   const filePath = getFilePath({ publicDir }, pagePath)
-  return fs.remove(filePath)
+  if (fs.existsSync(filePath)) {
+    return await fs.remove(filePath)
+  }
+  return Promise.resolve()
 }
 
 const write = async ({ publicDir }, page, result) => {
